feat: create git creds modal seperated
