+++ src/views/pages/GitCredentialManagement.vue
@@ -1,68 +1,21 @@
 <script setup>
 import PageBar from '@/views/components/PageBar.vue'
 import FilledButton from '@/views/components/FilledButton.vue'
-import ModalDialog from '@/views/components/ModalDialog.vue'
 import { useToast } from 'vue-toastification'
-import { computed, reactive, ref } from 'vue'
+import { computed, ref } from 'vue'
 import Table from '@/views/components/Table/Table.vue'
 import TableHeader from '@/views/components/Table/TableHeader.vue'
 import gql from 'graphql-tag'
 import { useMutation, useQuery } from '@vue/apollo-composable'
 import GitCredentialListRow from '@/views/partials/GitCredentialListRow.vue'
 import TableMessage from '@/views/components/Table/TableMessage.vue'
-import { preventSpaceInput } from '@/vendor/utils.js'
+import CreateGitCredentialModal from '@/views/partials/CreateGitCredentialModal.vue'
 
 const toast = useToast()
-const isModalOpen = ref(false)
-const openModal = () => {
-  isModalOpen.value = true
-}
-const closeModal = () => {
-  isModalOpen.value = false
-}
 
-// New Git Credential form state
-const newGitCredential = reactive({
-  name: '',
-  username: '',
-  password: ''
-})
-
-const {
-  mutate: createGitCredential,
-  loading: isGitCredentialCreating,
-  onDone: onGitCredentialCreateSuccess,
-  onError: onGitCredentialCreateFail
-} = useMutation(
-  gql`
-    mutation ($input: GitCredentialInput!) {
-      createGitCredential(input: $input) {
-        id
-        name
-        username
-        password
-      }
-    }
-  `,
-  {
-    variables: {
-      input: newGitCredential
-    }
-  }
-)
-
-onGitCredentialCreateSuccess(() => {
-  closeModal()
-  refetchGitCredentialList()
-  newGitCredential.name = ''
-  newGitCredential.username = ''
-  newGitCredential.password = ''
-  toast.success('Git Credential created successfully')
-})
-
-onGitCredentialCreateFail((err) => {
-  toast.error(err.message)
-})
+// Create Git Credential
+const createGitCredentialModalRef = ref(null)
+const openCreateGitCredentialModal = computed(() => createGitCredentialModalRef.value?.openModal ?? (() => {}))
 
 // Delete Git Credential mutation
 const {
@@ -132,71 +85,13 @@ onGitCredentialListError((err) => {
 <template>
   <section class="mx-auto w-full max-w-7xl">
     <!-- Modal for create -->
-    <ModalDialog :close-modal="closeModal" :is-open="isModalOpen">
-      <template v-slot:header>Add Git Credential</template>
-      <template v-slot:body>
-        Enter the necessary information for configuring the new Git Credential.
-        <form @submit.prevent="">
-          <!--  Name Field   -->
-          <div class="mt-4">
-            <label class="block text-sm font-medium text-gray-700" for="name">
-              Name (Provide a name to identify the credential)
-            </label>
-            <div class="mt-1">
-              <input
-                id="name"
-                v-model="newGitCredential.name"
-                autocomplete="off"
-                class="block w-full rounded-md border-gray-300 shadow-sm focus:border-primary-500 focus:ring-primary-500 sm:text-sm"
-                name="name"
-                placeholder="Name"
-                type="text" />
-            </div>
-          </div>
-          <!-- Username Field -->
-          <div class="mt-4">
-            <label class="block text-sm font-medium text-gray-700" for="username"> Git Username </label>
-            <div class="mt-1">
-              <input
-                id="username"
-                v-model="newGitCredential.username"
-                @keydown="preventSpaceInput"
-                autocomplete="off"
-                class="block w-full rounded-md border-gray-300 shadow-sm focus:border-primary-500 focus:ring-primary-500 sm:text-sm"
-                name="username"
-                placeholder="Git Username"
-                type="text" />
-            </div>
-          </div>
-          <!-- Password Field -->
-          <div class="mt-4">
-            <label class="block text-sm font-medium text-gray-700" for="password"> Git Password </label>
-            <div class="mt-1">
-              <input
-                id="password"
-                v-model="newGitCredential.password"
-                autocomplete="new-password"
-                class="block w-full rounded-md border-gray-300 shadow-sm focus:border-primary-500 focus:ring-primary-500 sm:text-sm"
-                name="password"
-                placeholder="Git Password"
-                type="text" />
-            </div>
-          </div>
-        </form>
-      </template>
-      <template v-slot:footer>
-        <FilledButton :click="createGitCredential" :loading="isGitCredentialCreating" type="primary"
-          >Add Now
-        </FilledButton>
-      </template>
-    </ModalDialog>
-
+    <CreateGitCredentialModal ref="createGitCredentialModalRef" :callback-on-create="refetchGitCredentialList" />
     <!-- Top Page bar   -->
     <PageBar>
       <template v-slot:title>Git Credentials</template>
       <template v-slot:subtitle> Manage Git Credentials and usage in deployments</template>
       <template v-slot:buttons>
-        <FilledButton :click="openModal" type="primary">Add New</FilledButton>
+        <FilledButton :click="openCreateGitCredentialModal" type="primary">Add New</FilledButton>
       </template>
     </PageBar>
